Immortalize and initialize global variables correctly.

+++ compiler-rt/test/asan/TestCases/alloca_detect_custom_size_.cpp
@@ -4,22 +4,23 @@
 
 #include "defines.h"
 #include <assert.h>
+#include <cstdint>
 struct A {
   char a[3];
   int b[3];
 };
 
 ATTRIBUTE_NOINLINE void foo(int index, int len) {
-ATTRIBUTE_ALIGNED(32)
+  ATTRIBUTE_ALIGNED(32)
 #ifdef MSVC
   volatile struct A *str = (volatile struct A *)_alloca(len * sizeof(struct A));
 #else
   volatile struct A str[len];
 #endif
-  assert(!(reinterpret_cast<long>(str) & 31L));
+  assert(!(reinterpret_cast<uintptr_t>(str) & 31L));
   str[index].a[0] = '1'; // BOOM
-// CHECK: ERROR: AddressSanitizer: dynamic-stack-buffer-overflow on address [[ADDR:0x[0-9a-f]+]]
-// CHECK: WRITE of size 1 at [[ADDR]] thread T0
+  // CHECK: ERROR: AddressSanitizer: dynamic-stack-buffer-overflow on address [[ADDR:0x[0-9a-f]+]]
+  // CHECK: WRITE of size 1 at [[ADDR]] thread T0
 }
 
 int main(int argc, char **argv) {
